requirements.txt
pandas
scikit-learn
fastapi[all]
httpx
uvicorn[standard]
python-dotenv